Download and install the libstdc++ Doxygen tags
* tags/mm-common-libstdc++.pc.in: New pkg-config data file with meta
information about the libstdc++ Doxygen documentation, in particular
where to find the tag file.
* tags/mm-common-libstdc++-uninstalled.pc.in: Likewise, but for use
without installing.
* configure.ac: Check for the presence of the curl and wget utilities,
and make it mandatory in maintainer mode that at least one of them is
available.  Add the pkg-config files to the output file list.
* Makefile.am (tags/libstdc++.tag): New custom rule for downloading
the libstdc++ tag file from the gcc.gnu.org website.  Install the
pkg-config data files so that dependent modules can find the tag file
and the location of the libstdc++ online documentation.