Make mm-common-prepare copy doc tools on request
* Makefile.am (dist_doctool_DATA): Install documentation utilities
to the $(pkgdatadir)/doctool directory.
* macros/mm-doc.m4 (MM_CONFIG_DOCTOOL_DIR): New Autoconf macro for
locating the documentation utilities.  Either a local directory in
the source tree can be used, or alternatively pkg-config will be
invoked and request the install location from the glibmm-2.4 module.
* util/mm-common-prepare.in: Scan configure.ac for calls to the new
MM_CONFIG_DOCTOOL_DIR() macro.  If a local directory name was given,
copy the documentation utilities into the source tree.