Restructure and comment MM_ARG_WITH_TAGFILE_DOC
* macros/mm-doc.m4 (_MM_ARG_WITH_TAGFILE_DOC): Add a fourth parameter
pkg-variable to avoid repeating the M4 expression to select between
the ${htmlrefdir} and ${htmlrefpub} values for @DOXYGEN_TAGFILES@.
Simplify the AS_CASE logic used to test for a trailing slash.  Move
the code which appends a trailing slash to ${htmlrefpub} to a later
point in the shell execution, in order to correctly append the slash
to a user-supplied local path as well.
(MM_ARG_WITH_TAGFILE_DOC): Pass the additional pkg-variable argument
to _MM_ARG_WITH_TAGFILE_DOC().