Do not append slash to DOXYGEN_TAGFILES locations
* build/doc-reference.am (dist_devhelpDATA_INSTALL): Strip trailing
slash from $(htmlrefdir), if there is one.
(dh_xsl_params): Strip trailing slash from $(htmlrefpub).
* macros/mm-doc.m4 (_MM_ARG_WITH_TAGFILE_DOC): Strip the trailing
slash from $mm_htmlrefdir and $mm_htmlrefpub if there is one, instead
of appending one.  Do append the slash to the locations listed in the
DOCINSTALL_FLAGS output variable, but omit it for DOXYGEN_TAGFILES.
Apparently Doxygen is not trying to be smart when it automatically
appends a trailing slash to the external documentation base paths.
Rather, it just mindlessly appends it no matter what, ending up with
a double slash if there already was one.