Refine wrap_init.cc and documentation dependencies
* build/compile-binding.am (binding_libprefix): Replace short-hand
substitution $(binding_name:lib%=) with an explicit $(filter) call.
* build/doc-reference.am (doc_dependencies): In maintainer mode, also
depend on Makefile.in in order to trigger a rebuild when the list of
input files changed.
* build/generate-binding.am (binding_stampfiles): Replace short-hand
substitution $(files_hg:%.hg=...) with an explicit $(patsubst) call.
(wrap_init.cc): Make Makefile.in a prerequisite so that the file will
be rebuilt when the list of source files changed.