Generate a nested hierarchy of Devhelp chapters
* util/tagfile-to-devhelp2.xsl (chapters): Turn the flat structure of
the Doxygen tag file into hierarchies of modules and classes.  At the
top-level, select only those compounds which are not a member of any
other compound, and then process the nested compounds recursively to
create the tree structure.
(functions): Refactor and canonicalize the template logic.  No longer
attempt to sort the keyword list, because it is not a requirement and
unnecessarily complicates the transformation.
(xsl:strip-space): Strip excess whitespace from source elements.