Process documentation targets conditionally
* build/doc-reference.am (dist_reference_DATA), (dist_noinst_DATA):
Assign the list of files to install and distribute conditionally,
depending on whether ENABLE_DOCUMENTATION is true.  Thus, it is no
longer required to skip the entire documentation build directory if
the documentation is disabled.
(install-data-local), (uninstall-data): As above, but for the phony
targets to install and uninstall files to the system.
($(doc_config)): Do not add any commands to the rule when building
from the top-level directory.  Automake automatically generates rules
and commands to rebuild config.status output files at the top level.