Add mm-common-util.pc so we can get doc utils from mm-common.
* util/mm-common-util.pc.in: Added this file, with the path to
the installed doctool/*.pl files.
* configure.ac:
* Makefile.am: Generate and install the .pc file.
* macros/mm-doc.m4: Use the mm-common-util .pc file, not the glibmm-2.4 one.

Previously, glibmm installed a copy that it got from mm-common-prepare
and other modules then asked glibmm where to find its copy.
That was probably for backwards compatibility.
But we break stuff with gtkmm 3 anyway, so now is a good time to
avoid the duplication by requiring modules to get it from mm-common instead.
Actually this should only require a rerun of autogen.sh, because
mm-common's own MM_CONFIG_DOCTOOL_DIR macro is what gets the path to
these utils.

Removing the hard-coding of glibmm-2.4 will also help us when we eventually
break the glibmm ABI.