Rename directory build/ to am_include/
This is in preparation for building mm-common with meson in the near future.

When jhbuild is used for building a module with meson, and the user has
requested builddir==srcdir, which meson forbids, jhbuild creates
a $srcdir/build directory and builds there, even if there already is such
a directory, used for other purposes.